+++ src/optimum/rbln/transformers/models/llava_next/modeling_llava_next.py
@@ -332,7 +332,9 @@ class RBLNLlavaNextForConditionalGeneration(RBLNModel, RBLNDecoderOnlyGeneration
             # otherwise has to be stacked from list of (num_patches, num_channels, height, width)
             raise ValueError(f"pixel_values of shape {pixel_values.shape}, expect to be of 4 or 5 dimensions")
 
-        image_features = self.vision_tower(pixel_values, output_hidden_states=True, out=vision_out_buffer)
+        image_features = self.vision_tower(
+            pixel_values.to(self.rbln_config.vision_tower.dtype), output_hidden_states=True, out=vision_out_buffer
+        )
         # If we have one vision feature layer, return the corresponding hidden states,
         # otherwise, select the hidden states of each feature layer and concatenate them
         if isinstance(vision_feature_layer, int):
